Скорости соединения теперь имеют тип unsigned log (или uint32_t). Все устройства подключения теперь принимают свои скорости в конструкторе.

+++ code/Arduino/connectors/ConnectingDevice.h
@@ -9,7 +9,7 @@ protected:
 
 	Stream *device;
 public:
-	ConnectingDevice(int rx, int tx, int speed);
+	ConnectingDevice(int rx, int tx, unsigned long speed);
 	virtual bool isActive();
 	virtual String read();
 	virtual void send(String data);
